Delete images/image.png
This is a placeholder file that is used to store the image used for the embed for server checks. Currently, it just alternates between the pass/fail image, but in the future, the intent is this image will pull from the image used by the server, which is why this existed at all. However, this is generated when the program is run, and shouldn't be included in the repo.